refactor: 상품구매정보 리팩토링
